articles detail 페이지 정리, 좋아요 비동기

+++ articles/views.py
@@ -1,8 +1,10 @@
+import json
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Review, Comment, Search, Photo
 from .forms import ReviewForm, CommentForm, PhotoForm
 from django.db.models import Q
 from django.core.paginator import Paginator
+from django.http import JsonResponse
 
 
 def index(request):
@@ -65,34 +67,38 @@ def detail(request, pk):
 def update(request, pk):
     review = Review.objects.get(pk=pk)
     photos = Photo.objects.filter(review_id=review.pk)
-    if request.method == "POST":
-        review_form = ReviewForm(request.POST, request.FILES, instance=review)
-        photo_form = PhotoForm(request.POST, request.FILES)
-        images = request.FILES.getlist("image")
-        for photo in photos:
-            if photo.image:
-                photo.delete()
-        if review_form.is_valid() and photo_form.is_valid():
-            review = review_form.save(commit=False)
-            if len(images):
-                for image in images:
-                    image_instance = Photo(review=review, image=image)
+
+    if review.user == request.user:
+        if request.method == "POST":
+            review_form = ReviewForm(request.POST, request.FILES, instance=review)
+            photo_form = PhotoForm(request.POST, request.FILES)
+            images = request.FILES.getlist("image")
+            for photo in photos:
+                if photo.image:
+                    photo.delete()
+            if review_form.is_valid() and photo_form.is_valid():
+                review = review_form.save(commit=False)
+                if len(images):
+                    for image in images:
+                        image_instance = Photo(review=review, image=image)
+                        review.save()
+                        image_instance.save()
+                else:
                     review.save()
-                    image_instance.save()
-            else:
-                review.save()
-            return redirect("articles:index")
-    else:
-        review_form = ReviewForm(instance=review)
-        if photos:
-            photo_form = PhotoForm(instance=photos[0])
+                return redirect("articles:index")
         else:
-            photo_form = PhotoForm()
-    context = {
-        "review_form": review_form,
-        "photo_form": photo_form,
-    }
-    return render(request, "articles/create.html", context)
+            review_form = ReviewForm(instance=review)
+            if photos:
+                photo_form = PhotoForm(instance=photos[0])
+            else:
+                photo_form = PhotoForm()
+        context = {
+            "review_form": review_form,
+            "photo_form": photo_form,
+        }
+        return render(request, "articles/create.html", context)
+    else:
+        return redirect('accounts:wrong_approach')
 
 
 def search(request):
@@ -165,3 +171,19 @@ def map(request):
     return render(request,"articles/map.html")
 def map2(request):
     return render(request,"articles/map2.html")
+
+def like(request,pk):
+    review = Review.objects.get(pk = pk)
+    if request.user not in review.like_users.all():
+        review.like_users.add(request.user)
+        is_like = True
+    else:
+        review.like_users.remove(request.user)
+        is_like = False
+
+    data = {
+        'isLike' : is_like,
+        'likeCount' : review.like_users.count(),
+    }
+    
+    return JsonResponse(data)
